Machine Learning package for Gregory AI
